requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
colorama==0.4.6
eval_type_backport==0.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jsonpath-python==1.0.6
mistralai==1.2.3
mypy-extensions==1.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
